// interfaces/lcu-types.d.ts
export interface LolTftLolTftBackgrounds {
	backgrounds: Record<string, string>
}
// ...
export interface LolTftPassTftBattlepass {
	/** @format int32 */
	totalPointsEarned: number
	milestones: LolTftPassTftBattlepassMilestone[]
	bonuses: LolTftPassTftBattlepassMilestone[]
	activeMilestone: LolTftPassTftBattlepassMilestone
	info: LolTftPassTftBattlepassInfo
	/** @format int32 */
	lastViewedProgress: number
	lastViewedMilestone: LolTftPassTftBattlepassMilestone
	/** @format int32 */
	currentLevel: number
}

export interface LolTftPassTftBattlepassInfo {
	title: string
	description: string
	/** @format uint64 */
	startDate: number
	/** @format uint64 */
	endDate: number
	premium: boolean
	premiumTitle: string
	premiumEntitlementId: string
	pcPurchaseRequirement: string
	passId: string
	media: Record<string, string>
}

export interface LolTftPassTftBattlepassMilestone {
	milestoneId: string
	title: string
	description: string
	status: string
	/** @format int32 */
	pointsNeededForMilestone: number
	/** @format int32 */
	pointsEarnedForMilestone: number
	/** @format int32 */
	totalPointsForMilestone: number
	/** @format int32 */
	level: number
	iconImageUrl: string
	iconNeedsFrame: boolean
	rewards: LolTftPassTftBattlepassReward[]
	isPaid: boolean
	isLocked: boolean
	isKeystone: boolean
	isBonus: boolean
	isClaimRequestPending: boolean
}

export interface LolTftPassTftBattlepassReward {
	description: string
	iconUrl: string
	iconNeedsFrame: boolean
}